ai-service/requirements.txt
mem0ai 
langchain-openai
langchain
langchain_core
langchain_qdrant
langchain_google_genai
langchain_community
langchain_text_splitters
python-dotenv
langgraph
fastapi
uvicorn
pydantic